radius_xlat should return size_t

+++ src/main/radmin.c
@@ -94,9 +94,9 @@ int debug_flag = 0;
 struct main_config_t mainconfig;
 char *request_log_file = NULL;
 char *debug_log_file = NULL;
-int radius_xlat(UNUSED char *out, UNUSED int outlen, UNUSED const char *fmt,
-		UNUSED REQUEST *request,
-		UNUSED RADIUS_ESCAPE_STRING func, UNUSED void *arg)
+size_t radius_xlat(UNUSED char *out, UNUSED int outlen, UNUSED const char *fmt,
+		   UNUSED REQUEST *request,
+		   UNUSED RADIUS_ESCAPE_STRING func, UNUSED void *arg)
 {
 	return -1;
 }
